Fix the bug of qtest occurs with command `help`
Two parameters (`simulation` and `echo`) use type `bool` variables as
their values, but addresses of them are cast into `int*` when parameters
are added into `param_list` in function `init_cmd` (line 104 and 107).

At line 307 in console.c, that pointer is dereferenced as `int*`,
in order to be formatted as output, result in reading an `int` at a
memory location allocated to a `bool`.

Since the elements in `param_list` is sorted in lexicographic order, the
address sanitizer would catch the error of variable `echo` first, even
if `simulation` is added into the list before it.

To deal with the issue above, just change that two variables into `int`.